Let every plugin collect the data it needs.
