+++ lib/plugins/html/htmlBuilder.js
@@ -13,6 +13,9 @@ const fs = require('fs'),
   metricHelper = require('./metricHelper'),
   isEmpty = require('lodash.isempty');
 
+const summaryBoxesSetup = require('./setup/summaryBoxes'),
+  detailedSetup = require('./setup/detailed');
+
 Promise.promisifyAll(fs);
 
 const TIME_FORMAT = 'YYYY-MM-DD HH:mm:ss';
@@ -22,20 +25,18 @@ class HTMLBuilder {
     this.storageManager = context.storageManager;
     this.timestamp = context.timestamp.format(TIME_FORMAT);
     this.options = options;
-    this.dataCollection = context.dataCollection;
     this.summary = {};
     this.budget = context.budget;
     this.context = context;
   }
 
-  render() {
+  render(dataCollector) {
     const options = this.options;
-    const dataCollection = this.dataCollection;
     const name = this.context.name;
     const timestamp = this.timestamp;
-    const nTestedPages = Object.keys(dataCollection.urlPages).length;
+    const nTestedPages = Object.keys(dataCollector.getURLs()).length;
     log.debug('Render HTML for %s page(s) ', nTestedPages);
-    const errors = dataCollection.getErrorPages();
+    const errors = dataCollector.getErrorUrls();
 
     if (Object.keys(errors).length > 0) {
       this.summary.errors = {
@@ -44,9 +45,9 @@ class HTMLBuilder {
       };
     }
 
-    const validPages = dataCollection.getValidPages();
-    const summaryBoxes = dataCollection.getSummaryBoxes();
-    const detailedBoxes = dataCollection.getDetailedBoxes();
+    const validPages = dataCollector.getWorkingUrls();
+    const summaryBoxes = summaryBoxesSetup(dataCollector.getSummary()['index']);
+    const detailedBoxes = detailedSetup(dataCollector.getSummary()['detailed']);
 
     this.summary.pages = {
       pageTitle: `Overview of ${helpers.plural(
@@ -126,9 +127,9 @@ class HTMLBuilder {
         merge(
           {
             options,
-            noPages: Object.keys(dataCollection.urlPages).length
+            noPages: Object.keys(dataCollector.getURLs()).length
           },
-          dataCollection.summaryPages[name],
+          dataCollector.getSummary()[name],
           this.summary[name]
         )
       )
@@ -136,7 +137,7 @@ class HTMLBuilder {
 
     const urlPageRenders = Promise.resolve(Object.keys(validPages)).map(url => {
       const pageInfo = validPages[url];
-      const runPages = dataCollection.urlRunPages[url];
+      const runPages = dataCollector.getURLRun()[url];
       const medianRun = metricHelper.pickMedianRun(runPages, pageInfo);
       let summaryPageHAR = get(pageInfo, 'data.browsertime.har');
       // if we don't use Browsertime, we don't get the browser version
@@ -189,13 +190,14 @@ class HTMLBuilder {
         )
       );
     });
-
     // Aggregate/summarize data and write additional files
     return Promise.all(summaryRenders)
       .then(() => Promise.all(urlPageRenders))
-      .then(() =>
-        this.storageManager.copyToResultDir(path.join(__dirname, 'assets'))
-      )
+      .then(() => {
+        return this.storageManager.copyToResultDir(
+          path.join(__dirname, 'assets')
+        );
+      })
       .then(() =>
         log.info('HTML stored in %s', this.storageManager.getBaseDir())
       );
