+++ lib/plugins/html/setup/detailed.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const h = require('../helpers');
+const h = require('../../../support/helpers');
 const get = require('lodash.get');
 
 function row(stat, name, metricName, formatter) {
