+++ lib/plugins/html/setup/summaryBoxes.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const h = require('../helpers');
+const h = require('../../../support/helpers');
 const get = require('lodash.get');
 
 function infoBox(stat, name, formatter, url) {
