+++ lib/plugins/slack/attachements.js
@@ -11,10 +11,10 @@ function getMetric(metric) {
   }
 }
 
-module.exports = function(dataCollection, resultUrls, slackOptions) {
+module.exports = function(dataCollector, resultUrls, slackOptions) {
   const attachments = [];
 
-  const resultsPerUrl = dataCollection.urlPages;
+  const resultsPerUrl = dataCollector.getURLs();
 
   for (let url of Object.keys(resultsPerUrl)) {
     const results = resultsPerUrl[url];
