+++ lib/plugins/slack/index.js
@@ -5,6 +5,8 @@ const Promise = require('bluebird');
 const log = require('intel').getLogger('sitespeedio.plugin.slack');
 const Slack = require('node-slack');
 const merge = require('lodash.merge');
+const set = require('lodash.set');
+const DataCollector = require('../../support/dataCollector');
 const getAttachments = require('./attachements');
 const getSummary = require('./summary');
 
@@ -21,15 +23,50 @@ const defaultConfig = {
 module.exports = {
   open(context, options) {
     throwIfMissing(options.slack, ['hookUrl', 'userName'], 'slack');
-    this.data = context.dataCollection;
+    this.dataCollector = new DataCollector(context);
     this.resultUrls = context.resultUrls;
     this.context = context;
+  },
+  processMessage(message) {
+    const dataCollector = this.dataCollector;
+
+    switch (message.type) {
+      case 'error': {
+        dataCollector.addErrorForUrl(message.url, message.source, message.data);
+        break;
+      }
+
+      case 'browsertime.run':
+      case 'browsertime.pageSummary':
+      case 'pagexray.run':
+      case 'pagexray.pageSummary':
+      case 'coach.run':
+      case 'coach.pageSummary': {
+        dataCollector.addDataForUrl(
+          message.url,
+          message.type,
+          message.data,
+          message.runIndex
+        );
+        break;
+      }
+
+      case 'coach.summary':
+      case 'pagexray.summary':
+      case 'browsertime.summary': {
+        const data = {};
+        set(data, message.type, message.data);
+        dataCollector.addSummary('index', data);
+        break;
+      }
+    }
   },
   final(options) {
     const slackOptions = merge({}, defaultConfig, options.slack);
     const slack = new Slack(slackOptions.hookUrl);
     const type = slackOptions.type;
     const pageErrors = [];
+    const dataCollector = this.dataCollector;
     let logo = 'https://www.sitespeed.io/img/slack/sitespeed-logo-slack.png';
 
     let channel = slackOptions.channel;
@@ -37,9 +74,9 @@ module.exports = {
       channel = `#${channel}`;
     }
 
-    for (const url of Object.keys(this.data.urlPages)) {
-      if (this.data.urlPages[url].errors) {
-        pageErrors.push(this.data.urlPages[url].errors);
+    for (const url of Object.keys(dataCollector.getURLs())) {
+      if (dataCollector.getURLs()[url].errors) {
+        pageErrors.push(dataCollector.getURLs()[url].errors);
       }
     }
 
@@ -47,7 +84,7 @@ module.exports = {
 
     if (['summary', 'all', 'error'].includes(type)) {
       const sum = getSummary(
-        this.data,
+        dataCollector,
         pageErrors,
         this.resultUrls,
         this.context,
@@ -59,7 +96,11 @@ module.exports = {
 
     let attachments = [];
     if (['url', 'all', 'error'].includes(type)) {
-      attachments = getAttachments(this.data, this.resultUrls, slackOptions);
+      attachments = getAttachments(
+        dataCollector,
+        this.resultUrls,
+        slackOptions
+      );
     }
 
     if ((type === 'error' && pageErrors.length > 0) || type !== 'error') {
