+++ lib/plugins/slack/summary.js
@@ -3,14 +3,8 @@ const util = require('util');
 const get = require('lodash.get');
 const h = require('../../support/helpers');
 
-module.exports = function(
-  dataCollection,
-  errors,
-  resultUrls,
-  context,
-  options
-) {
-  const base = dataCollection.summaryPages.index || {};
+module.exports = function(dataCollector, errors, resultUrls, context, options) {
+  const base = dataCollector.getSummary().index || {};
   const metrics = {
     firstPaint: {
       name: 'First paint',
@@ -64,7 +58,7 @@ module.exports = function(
 
   let summaryText =
     `${h.plural(
-      Object.keys(dataCollection.urlPages).length,
+      Object.keys(dataCollector.getURLs()).length,
       'page'
     )} analyzed for ${h.short(context.name, 30)} ` +
     `(${h.plural(options.browsertime.iterations, 'run')}, ` +
