+++ lib/plugins/text/textBuilder.js
@@ -5,6 +5,7 @@ const table = require('text-table'),
   color = require('cli-color'),
   chunk = require('lodash.chunk'),
   h = require('../../support/helpers'),
+  summaryBoxesSetup = require('../html/setup/summaryBoxes'),
   tableOpts = {
     stringLength: color.getStrippedLength
   },
@@ -21,8 +22,8 @@ function getColor(label) {
   ];
 }
 
-function getHeader(dataCollection, context, options) {
-  const noPages = Object.keys(dataCollection.urlPages).length;
+function getHeader(dataCollector, context, options) {
+  const noPages = options.urls.length;
   return drab(
     [
       `${h.plural(noPages, 'page')} analyzed for ${h.short(context.name, 30)} `,
@@ -34,8 +35,13 @@ function getHeader(dataCollection, context, options) {
   );
 }
 
-function getBoxes(dataCollection) {
-  return flatten(chunk(dataCollection.getSummaryBoxes().filter(Boolean), 3));
+function getBoxes(dataCollector) {
+  return flatten(
+    chunk(
+      summaryBoxesSetup(dataCollector.getSummary()['text']).filter(Boolean),
+      3
+    )
+  );
 }
 
 // foo bar -> fb
@@ -49,9 +55,9 @@ function noop(a) {
 }
 
 module.exports = {
-  renderSummary(dataCollection, context, options) {
-    let out = getHeader(dataCollection, context, options);
-    let rows = getBoxes(dataCollection).map(b => {
+  renderSummary(dataCollector, context, options) {
+    let out = getHeader(dataCollector, context, options);
+    let rows = getBoxes(dataCollector).map(b => {
       var marker = getMarker(b.label),
         c = getColor(b.label);
       // color.xterm(ansi)(label),
@@ -64,14 +70,14 @@ module.exports = {
     options.summary = { out: `${out}\n` + table(rows, tableOpts) };
   },
 
-  renderBriefSummary(dataCollection, context, options) {
-    let out = getHeader(dataCollection, context, options);
+  renderBriefSummary(dataCollector, context, options) {
+    let out = getHeader(dataCollector, context, options);
     var lines = [],
       scores = [],
       size = '',
       reqs = '',
       rum = '';
-    getBoxes(dataCollection).map(b => {
+    getBoxes(dataCollector).map(b => {
       var c = getColor(b.label),
         val = b.median,
         name;
