+++ lib/sitespeed.js
@@ -11,7 +11,6 @@ const Promise = require('bluebird'),
   merge = require('lodash.merge'),
   pullAll = require('lodash.pullall'),
   union = require('lodash.union'),
-  DataCollection = require('./support/dataCollection'),
   browsertimeConfig = require('./plugins/browsertime').config,
   webpagetestConfig = require('./plugins/webpagetest').config,
   packageInfo = require('../package');
@@ -60,8 +59,6 @@ module.exports = {
       options.outputFolder,
       options.resultBaseURL
     );
-
-    const dataCollection = new DataCollection();
 
     return storageManager
       .createDataDir('logs')
@@ -138,7 +135,6 @@ module.exports = {
             {
               storageManager,
               resultUrls,
-              dataCollection,
               timestamp,
               budget: budgetResult,
               name: url,
