+++ lib/support/dataCollector.js
@@ -4,6 +4,7 @@ const fs = require('fs'),
   Promise = require('bluebird'),
   merge = require('lodash.merge'),
   get = require('lodash.get'),
+  reduce = require('lodash.reduce'),
   set = require('lodash.set');
 
 Promise.promisifyAll(fs);
@@ -11,9 +12,9 @@ Promise.promisifyAll(fs);
 class DataCollector {
   constructor(context) {
     this.resultUrls = context.resultUrls;
-    this.urlRunPages = context.dataCollection.urlRunPages;
-    this.urlPages = context.dataCollection.urlPages;
-    this.summaryPages = context.dataCollection.summaryPages;
+    this.urlRunPages = {};
+    this.urlPages = {};
+    this.summaryPages = {};
   }
 
   _addUrl(url) {
@@ -24,6 +25,44 @@ class DataCollector {
       };
       this.urlRunPages[url] = [];
     }
+  }
+
+  getSummary() {
+    return this.summaryPages;
+  }
+
+  getURLRun() {
+    return this.urlRunPages;
+  }
+
+  getURLs() {
+    return this.urlPages;
+  }
+
+  getWorkingUrls() {
+    return reduce(
+      this.urlPages,
+      (validPages, urlInfo, url) => {
+        if (Object.keys(urlInfo.data).length > 0) {
+          validPages[url] = urlInfo;
+        }
+        return validPages;
+      },
+      {}
+    );
+  }
+
+  getErrorUrls() {
+    return reduce(
+      this.urlPages,
+      (errors, urlInfo, url) => {
+        if (urlInfo.errors) {
+          errors[url] = urlInfo.errors;
+        }
+        return errors;
+      },
+      {}
+    );
   }
 
   addErrorForUrl(url, source, data) {
@@ -49,7 +88,7 @@ class DataCollector {
     }
   }
 
-  addDataForSummaryPage(name, data) {
+  addSummary(name, data) {
     if (this.summaryPages[name]) {
       merge(this.summaryPages[name], data);
     } else {
