+++ lib/support/pluginLoader.js
@@ -9,7 +9,6 @@ Promise.promisifyAll(fs);
 const defaultPlugins = new Set([
   'browsertime',
   'coach',
-  'datacollector',
   'domains',
   'assets',
   'html',
